Bunch of changes related to redfox. Too many to describe...

+++ modules/Phlex/RedFox/Model/Field/FloatField.php
@@ -2,7 +2,7 @@
 
 use Phlex\RedFox\Model\Field;
 
-class IntegerField extends Field{
+class FloatField extends Field{
 	protected $min;
 	protected $max;
